requirements.txt
asgiref==3.7.2
Django==4.2.7
sqlparse==0.4.4
psycopg[binary]==3.1.12
django-environ==0.10.0
djangorestframework
drf-yasg
Pillow